# client/src/api/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Menu+ API",
                 "version":  "0.1.0"
             },
    "paths":  {
                  "/auth/register":  {
                                         "post":  {
                                                      "tags":  [
                                                                   "Auth"
                                                               ],
                                                      "summary":  "Register",
                                                      "operationId":  "register_auth_register_post",
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/Register"
                                                                                                                              }
                                                                                                               }
                                                                                      },
                                                                          "required":  true
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Successful Response",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {

                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                        "422":  {
                                                                                    "description":  "Validation Error",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                }
                                                                    }
                                                  }
                                     },
                  "/auth/login":  {
                                      "post":  {
                                                   "tags":  [
                                                                "Auth"
                                                            ],
                                                   "summary":  "Sign In",
                                                   "operationId":  "sign_in_auth_login_post",
                                                   "requestBody":  {
                                                                       "content":  {
                                                                                       "application/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/Login"
                                                                                                                           }
                                                                                                            }
                                                                                   },
                                                                       "required":  true
                                                                   },
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "Successful Response",
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {

                                                                                                                                     }
                                                                                                                      }
                                                                                             }
                                                                             },
                                                                     "422":  {
                                                                                 "description":  "Validation Error",
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                     }
                                                                                                                      }
                                                                                             }
                                                                             }
                                                                 }
                                               }
                                  },
                  "/auth/verify-signup":  {
                                              "post":  {
                                                           "tags":  [
                                                                        "Auth"
                                                                    ],
                                                           "summary":  "Verify Signup",
                                                           "operationId":  "verify_signup_auth_verify_signup_post",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "email",
                                                                                  "in":  "query",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string",
                                                                                                 "title":  "Email"
                                                                                             }
                                                                              },
                                                                              {
                                                                                  "name":  "token",
                                                                                  "in":  "query",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string",
                                                                                                 "title":  "Token"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "Successful Response",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {

                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     },
                                                                             "422":  {
                                                                                         "description":  "Validation Error",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       }
                                          },
...
              },
    "components":  {
                       "schemas":  {
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "Login":  {
                                                     "properties":  {
                                                                        "email":  {
                                                                                      "type":  "string",
                                                                                      "format":  "email",
                                                                                      "title":  "Email",
                                                                                      "default":  "nazar.konechniy2@gmail.com"
                                                                                  },
                                                                        "password":  {
                                                                                         "type":  "string",
                                                                                         "title":  "Password",
                                                                                         "default":  "nazar.konechniy2@gmail.com"
                                                                                     }
                                                                    },
                                                     "type":  "object",
                                                     "title":  "Login"
                                                 },
                                       "Register":  {
                                                        "properties":  {
                                                                           "email":  {
                                                                                         "type":  "string",
                                                                                         "title":  "Email"
                                                                                     },
                                                                           "password":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Password"
                                                                                        }
                                                                       },
                                                        "type":  "object",
                                                        "required":  [
                                                                         "email",
                                                                         "password"
                                                                     ],
                                                        "title":  "Register"
                                                    },
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           },
...
                                   },
                       "securitySchemes":  {
                                               "APIKeyHeader":  {
                                                                    "type":  "apiKey",
                                                                    "in":  "header",
                                                                    "name":  "Authorization"
                                                                }
                                           }
                   }
}